requirements.txt
numpy==1.21.6
opencv-python==4.8.0.76
tqdm==4.66.1
transformers==4.30.2
imgaug==0.4.0
pandas==1.3.5
Pillow==9.5.0
seaborn==0.12.2
pycocotools==2.0.7
matplotlib==3.5.3
jupyterlab